Tests: rename apdu.py to backend.py
