+++ tests/integration/nano/test_ensure_always_clearsign.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from utils.apdu import *
 from utils.app import *
+from utils.backend import *
 
 # Operation (0): Transaction
 # Fee: 0.01 XTZ
