+++ tests/integration/nano/test_reject_pk.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from utils.apdu import *
 from utils.app import *
+from utils.backend import *
 
 if __name__ == "__main__":
     test_name = Path(__file__).stem
