+++ tests/integration/nano/test_sign_transfer_ticket.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from utils.apdu import *
 from utils.app import *
+from utils.backend import *
 
 # Operation (0): Transfer ticket
 # Fee: 0.01 XTZ
