+++ tests/integration/nano/test_tz3_sign_micheline_basic.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from utils.apdu import *
 from utils.app import *
+from utils.backend import *
 
 # Expression: {"CACA";"POPO";"BOUDIN"}
