+++ tests/integration/nano/test_wrong_apdu_length.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from utils.apdu import *
 from utils.app import *
+from utils.backend import *
 
 if __name__ == "__main__":
     with nano_app() as app:
